requirements.txt
flask
python-dotenv
twilio
requests
openai